requirements.txt
beautifulsoup4~=4.6.0
openpyxl~=2.4.9
Pillow~=4.3.0
requests~=2.18.4

